Remove HoldingQueue
Why:

* For v2 we don't want to use the `HoldingQueue`. Instead we want
notifications to go straight into the `SendingQueue`.
* Asana link: https://app.asana.com/0/529741067494252/620981192963538
