+++ apps/alert_processor/lib/dissemination/sms_opt_out_worker.ex
@@ -9,7 +9,6 @@ defmodule AlertProcessor.SmsOptOutWorker do
   require Logger
   alias AlertProcessor.Aws.AwsClient
   alias AlertProcessor.Model.User
-  alias AlertProcessor.HoldingQueue
   alias AlertProcessor.Helpers.ConfigHelper
 
   @type phone_number :: String.t
@@ -106,6 +105,5 @@ defmodule AlertProcessor.SmsOptOutWorker do
   @spec update_users_opted_out([User.id]) :: [User.id]
   defp update_users_opted_out(user_ids) do
     User.remove_users_phone_number(user_ids, "sms-opt-out")
-    Enum.map(user_ids, &HoldingQueue.remove_user_notifications/1)
   end
 end
