+++ apps/alert_processor/test/alert_processor/dissemination/sending_queue_test.exs
@@ -37,4 +37,17 @@ defmodule AlertProcessor.SendingQueueTest do
     assert notification == notification2
     assert :error = SendingQueue.pop(test)
   end
+
+  test "Multiple alerts can be added to the queue", %{test: test} do
+    notification1 = %Notification{alert_id: "1"}
+    notification2 = %Notification{alert_id: "2"}
+    notification3 = %Notification{alert_id: "3"}
+    SendingQueue.start_link([name: test])
+    SendingQueue.list_enqueue(test, [notification1])
+    SendingQueue.list_enqueue(test, [notification2, notification3])
+
+    assert SendingQueue.pop(test) == {:ok, notification2}
+    assert SendingQueue.pop(test) == {:ok, notification3}
+    assert SendingQueue.pop(test) == {:ok, notification1}
+  end
 end
